Put amod in same folder as moac for AppImage
I had tested the AppImage build, but I forgot to test mods. It looks
like amod needs to be in the same directory as moac, and can't live in
/usr/lib (even though for AppImage that's usually where all dynamic libs
would live). This adjusts the directory structure so amod lives next to
moac. Tested and confirmed as working to support mods as intended on
Linux.