Switch to mimalloc for all prev malloc calls
This switches to mimalloc for all previous usages of malloc and free,
other than allocations where we defer to external libraries to alloc and
free for us.

In testing mimalloc resolved hitching issues on Windows, and it's a
better overall performer than the system allocators for Windows, OSX, or
Linux.