Update CI pipeline and docker builds for mimalloc
This updates our CI pipeline along with all docker builds for the use of
mimalloc. This has a nice benefit that we can bake in LD_PRELOAD for the
AppImage, so Linux users that use the AppImage will get a full global
override, ensuring 100% of our libraries use mimalloc on linux.